+++ frontend/src/AuthRoute.js
@@ -27,28 +27,24 @@ export function ReToken() {
     })
     .then((result) => {
         if (result.error_code !== null) {
-          alert(result.error_code, result.data.message);
+          alert(result.error_code, result.message);
           alert("로그인이 필요합니다.");
           window.location.href = "/login";
           return;
         }
+        console.log("reissue succeed");
       })
     .catch((error) => console.log("error", error));
 }
 
 function AuthRoute({ component: Component, render, ...rest }) {
-  console.log("진입");
   if (localStorage.getItem("authorization")) {
     var temp = localStorage.getItem("authorization");
-    var jwt_decode = require("jwt-decode");
-    var decodeToken;
-    console.log(temp)
+
     if (temp !== null) {
-      decodeToken = jwt_decode(temp);
+      let exp = localStorage.getItem("exp")
       // token 유효시간 체크
-      console.log(decodeToken.exp * 1000, Date.now(), decodeToken.exp * 1000 - Date.now());
-      if (decodeToken.exp * 1000 - Date.now() < 1000 * 30) {
-        console.log("re");
+      if (exp * 1000 - Date.now() < 1000 * 30) {
         ReToken();
       }
     }
